requirements.txt
langchain-openai
langchain
python-dotenv
langchain_core
langchain_community
streamlit
langchain-groq